using the parameter class in oep, moving class and implementation into apps/cem
