+++ src/apps/chem/CMakeLists.txt
@@ -10,13 +10,15 @@ set(MADCHEM_HEADERS
     molecular_optimizer.h projector.h
     SCFOperators.h CCStructures.h CalculationParameters.h
     electronic_correlation_factor.h cheminfo.h vibanal.h molopt.h TDHF.h CC2.h CCPotentials.h
-    pcm.h SCFProtocol.h AC.h pointgroupoperator.h pointgroupsymmetry.h PNO.h PNOF12Potentials.h PNOGuessFunctions.h PNOParameters.h PNOStructures.h PNOTensors.h)
+    pcm.h SCFProtocol.h AC.h pointgroupoperator.h pointgroupsymmetry.h PNO.h PNOF12Potentials.h 
+    PNOGuessFunctions.h PNOParameters.h PNOStructures.h PNOTensors.h oep.h)
 set(MADCHEM_SOURCES
     correlationfactor.cc molecule.cc molecularbasis.cc vibanal.cc
     corepotential.cc atomutil.cc lda.cc cheminfo.cc
     distpm.cc SCF.cc gth_pseudopotential.cc nemo.cc mp2.cc pcm.cc
     SCFOperators.cc TDHF.cc GuessFactory.cc CCStructures.cc CC2.cc CCPotentials.cc 
-    AC.cc pointgroupsymmetry.cc QCCalculationParametersBase.cc PNO.cpp PNOF12Potentials.cpp PNOGuessFunctions.cpp PNOParameters.cpp PNOStructures.cpp)
+    AC.cc pointgroupsymmetry.cc QCCalculationParametersBase.cc PNO.cpp PNOF12Potentials.cpp 
+    PNOGuessFunctions.cpp PNOParameters.cpp PNOStructures.cpp oep.cc)
 if(LIBXC_FOUND)
   list(APPEND MADCHEM_SOURCES xcfunctional_libxc.cc)
 else()
